#40 #16 Made it possible to use lasso tool for heatmap
Plotly does not currently support lasso tool on heatmaps, but it does on scatterplots. So I've made a figure with two traces: (1) heatmap, (2) scatterplot that has one point per integer coordinate within the heatmap. This does require the user to select the center of each square in the heatmap, so I plot these centers as small black dots so the user may understand what is actually selected